Plot Outliers for a Single Sample in a SingleCellExperiment

This function generates a plot for a specified sample within a SingleCellExperiment object, highlighting outliers based on a specified metric. The plot visualizes the metric of interest and indicates outliers with a distinct color.

Usage

plotOutliers(
  spe,
  sample_id = "sample_id",
  sample = unique(spe$sample_id)[1],
  metric = "detected",
  outliers = NULL,
  point_size = 1,
  colors = c("white", "black"),
  stroke = 1
)

Arguments

spe

A SpatialExperiment object containing the data to be plotted.

sample_id

A character string specifying the column name in colData(spe) that contains unique sample identifiers. Default is "sample_id".

sample

A character string or numeric value specifying the sample to be plotted. By default, it plots the first unique sample found in spe$sample_id.

metric

A character string specifying the metric to be visualized in the plot. This metric should be a column name in colData(spe).

outliers

A character string specifying the column name in colData(spe) that indicates whether a data point is considered an outlier. Default is NULL.

colors

A character vector specifying the colors to be used for the gradient scale. If length is 2, the gradient will be a single color gradient.

stroke

A numeric value specifying the border thickness for outlier points. Default is 1.

Value

The function returns a plot object created by make_escheR and modified with additional layers for visualizing the specified metric and outliers. The plot is not explicitly printed by the function and should be printed by the caller.

Examples

plot <- plotOutliers(spe)
#> Error in eval(expr, envir, enclos): object 'spe' not found
plot
#> new("standardGeneric", .Data = function (x, y, ...) 
#> standardGeneric("plot"), generic = "plot", package = "base", 
#>     group = list(), valueClass = character(0), signature = c("x", 
#>     "y"), default = new("derivedDefaultMethod", .Data = function (x, 
#>         y, ...) 
#>     UseMethod("plot"), target = new("signature", .Data = "ANY", 
#>         names = "x", package = "methods"), defined = new("signature", 
#>         .Data = "ANY", names = "x", package = "methods"), generic = "plot"), 
#>     skeleton = (new("derivedDefaultMethod", .Data = function (x, 
#>         y, ...) 
#>     UseMethod("plot"), target = new("signature", .Data = "ANY", 
#>         names = "x", package = "methods"), defined = new("signature", 
#>         .Data = "ANY", names = "x", package = "methods"), generic = "plot"))(x, 
#>         y, ...))
#> <environment: 0x5622f1b78508>
#> attr(,"generic")
#> [1] "plot"
#> attr(,"generic")attr(,"package")
#> [1] "base"
#> attr(,"package")
#> [1] "base"
#> attr(,"group")
#> list()
#> attr(,"valueClass")
#> character(0)
#> attr(,"signature")
#> [1] "x" "y"
#> attr(,"default")
#> Method Definition (Class "derivedDefaultMethod"):
#> 
#> function (x, y, ...) 
#> UseMethod("plot")
#> <bytecode: 0x5622f1b3e7b0>
#> <environment: namespace:base>
#> 
#> Signatures:
#>         x    
#> target  "ANY"
#> defined "ANY"
#> attr(,"skeleton")
#> (new("derivedDefaultMethod", .Data = function (x, y, ...) 
#> UseMethod("plot"), target = new("signature", .Data = "ANY", names = "x", 
#>     package = "methods"), defined = new("signature", .Data = "ANY", 
#>     names = "x", package = "methods"), generic = "plot"))(x, 
#>     y, ...)
#> attr(,"class")
#> [1] "standardGeneric"
#> attr(,"class")attr(,"package")
#> [1] "methods"