localVariance Function

This function does calculates the local variance based on kNN.

Usage

localVariance(
  spe,
  n_neighbors = 36,
  features = c("expr_chrM_ratio"),
  samples = "sample_id",
  log2 = FALSE,
  n_cores = 1,
  name = NULL
)

Arguments

spe

SpatialExperiment object with the following columns in colData: sample_id, sum_umi, sum_gene

n_neighbors

Number of nearest neighbors to use for variance calculation

features

Features to use for variance calculation

samples

Column in colData to use for sample ID

log2

Whether to log2 transform the features

n_cores

Number of cores to use for parallelization

name

Name of the new column to add to colData

Value

SpatialExperiment object with feature variance added to colData

Examples

library(SpotSweeper)
library(SpatialExperiment)

# load example data
spe <- STexampleData::Visium_humanDLPFC()
#> see ?STexampleData and browseVignettes('STexampleData') for documentation
#> loading from cache

# change from gene id to gene names
rownames(spe) <- rowData(spe)$gene_name

# show column data before SpotSweepR
colnames(colData(spe))
#> [1] "barcode_id"   "sample_id"    "in_tissue"    "array_row"    "array_col"   
#> [6] "ground_truth" "cell_count"  

# drop out-of-tissue spots
spe <- spe[, spe$in_tissue == 1]
spe <- spe[, !is.na(spe$ground_truth)]

# Identifying the mitochondrial transcripts in our SpatialExperiment.
is.mito <- rownames(spe)[grepl("^MT-", rownames(spe))]

# Calculating QC metrics for each spot using scuttle
spe<- scuttle::addPerCellQCMetrics(spe, subsets=list(Mito=is.mito))
colnames(colData(spe))
#>  [1] "barcode_id"            "sample_id"             "in_tissue"            
#>  [4] "array_row"             "array_col"             "ground_truth"         
#>  [7] "cell_count"            "sum"                   "detected"             
#> [10] "subsets_Mito_sum"      "subsets_Mito_detected" "subsets_Mito_percent" 
#> [13] "total"                

spe <- localVariance(spe,
                     features = "subsets_Mito_percent",
                     n_neighbors=36,
                     name="local_mito_variance_k36"
                     )